4.21*: bloglist expansion, step9

+++ part4/blog/controllers/blogs.js
@@ -6,14 +6,6 @@ const User = require('../models/user')
 
 
 
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
-
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog
       .find({})  
@@ -45,8 +37,23 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    } else {
+      const blog = await Blog.findById(request.params.id)
+
+      console.log('ffff ', blog)
+
+      if (blog.user.toString() == decodedToken.id) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } else {
+        response.status(403).end()
+      }
+    }
 })
 
 blogRouter.put('/:id',  (request, response, next) => {
@@ -68,6 +75,6 @@ blogRouter.put('/:id',  (request, response, next) => {
     // const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     // response.status(200).json(updatedBlog)
   
-  })
+})
 
 module.exports = blogRouter
